Refactor: 댓글 lazyloading 함수 분리
lazyloading 부분 utils로 분리
